requirements.txt
flask
PyPDF2
scikit-learn
pymongo
python-dotenv